revised pagelist after merge with master - finetuned column preview and restructured code into different components for better readabilty
