+++ src/www/js/Api/DataSchema/ApiDocuments.ts
@@ -19,10 +19,9 @@ export interface PageInfo {
   foliation: string;
   pageId: number;
   foliationIsSet: boolean;
-
-  isTranscribed: boolean; // Wichtig für die CSS-Klasse 'without-transcription'
-  thumbnailUrl: string;   // Für die Anzeige in der Liste
-  jpgUrl: string;         // Als Fallback oder für den großen Viewer
+  isTranscribed: boolean;
+  thumbnailUrl: string;
+  jpgUrl: string;
 }
 
 export interface DocInfo {
